pgs-sup-generator/requirements.txt
numpy
opencv_python
Pillow
pytesseract
